requirements.txt
python-telegram-bot[all]
sqlalchemy
psycopg2-binary